# backend/api/gen/siem/v1/dashboards_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: siem/v1/dashboards.proto

package siemv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationDashboardsGetDisagreements = "/siem.v1.Dashboards/GetDisagreements"
const OperationDashboardsGetFeedHealthPanel = "/siem.v1.Dashboards/GetFeedHealthPanel"
const OperationDashboardsGetOverview = "/siem.v1.Dashboards/GetOverview"
const OperationDashboardsGetRules = "/siem.v1.Dashboards/GetRules"
const OperationDashboardsGetSources = "/siem.v1.Dashboards/GetSources"

type DashboardsHTTPServer interface {
	GetDisagreements(context.Context, *DashboardRequest) (*DisagreementsPanel, error)
	GetFeedHealthPanel(context.Context, *DashboardRequest) (*FeedHealthPanel, error)
	GetOverview(context.Context, *DashboardRequest) (*OverviewPanel, error)
	GetRules(context.Context, *DashboardRequest) (*RulesPanel, error)
	GetSources(context.Context, *DashboardRequest) (*SourcesPanel, error)
}

func RegisterDashboardsHTTPServer(s *http.Server, srv DashboardsHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/dashboards/overview", _Dashboards_GetOverview0_HTTP_Handler(srv))
	r.GET("/api/v1/dashboards/rules", _Dashboards_GetRules0_HTTP_Handler(srv))
	r.GET("/api/v1/dashboards/sources", _Dashboards_GetSources0_HTTP_Handler(srv))
	r.GET("/api/v1/dashboards/disagreements", _Dashboards_GetDisagreements0_HTTP_Handler(srv))
	r.GET("/api/v1/dashboards/feed-health", _Dashboards_GetFeedHealthPanel0_HTTP_Handler(srv))
}

func _Dashboards_GetOverview0_HTTP_Handler(srv DashboardsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DashboardRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDashboardsGetOverview)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetOverview(ctx, req.(*DashboardRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*OverviewPanel)
		return ctx.Result(200, reply)
	}
}

func _Dashboards_GetRules0_HTTP_Handler(srv DashboardsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DashboardRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDashboardsGetRules)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetRules(ctx, req.(*DashboardRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*RulesPanel)
		return ctx.Result(200, reply)
	}
}

func _Dashboards_GetSources0_HTTP_Handler(srv DashboardsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DashboardRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDashboardsGetSources)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSources(ctx, req.(*DashboardRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SourcesPanel)
		return ctx.Result(200, reply)
	}
}

func _Dashboards_GetDisagreements0_HTTP_Handler(srv DashboardsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DashboardRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDashboardsGetDisagreements)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetDisagreements(ctx, req.(*DashboardRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DisagreementsPanel)
		return ctx.Result(200, reply)
	}
}

func _Dashboards_GetFeedHealthPanel0_HTTP_Handler(srv DashboardsHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DashboardRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDashboardsGetFeedHealthPanel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetFeedHealthPanel(ctx, req.(*DashboardRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*FeedHealthPanel)
		return ctx.Result(200, reply)
	}
}

type DashboardsHTTPClient interface {
	GetDisagreements(ctx context.Context, req *DashboardRequest, opts ...http.CallOption) (rsp *DisagreementsPanel, err error)
	GetFeedHealthPanel(ctx context.Context, req *DashboardRequest, opts ...http.CallOption) (rsp *FeedHealthPanel, err error)
	GetOverview(ctx context.Context, req *DashboardRequest, opts ...http.CallOption) (rsp *OverviewPanel, err error)
	GetRules(ctx context.Context, req *DashboardRequest, opts ...http.CallOption) (rsp *RulesPanel, err error)
	GetSources(ctx context.Context, req *DashboardRequest, opts ...http.CallOption) (rsp *SourcesPanel, err error)
}

type DashboardsHTTPClientImpl struct {
	cc *http.Client
}

func NewDashboardsHTTPClient(client *http.Client) DashboardsHTTPClient {
	return &DashboardsHTTPClientImpl{client}
}

func (c *DashboardsHTTPClientImpl) GetDisagreements(ctx context.Context, in *DashboardRequest, opts ...http.CallOption) (*DisagreementsPanel, error) {
	var out DisagreementsPanel
	pattern := "/api/v1/dashboards/disagreements"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDashboardsGetDisagreements))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DashboardsHTTPClientImpl) GetFeedHealthPanel(ctx context.Context, in *DashboardRequest, opts ...http.CallOption) (*FeedHealthPanel, error) {
	var out FeedHealthPanel
	pattern := "/api/v1/dashboards/feed-health"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDashboardsGetFeedHealthPanel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DashboardsHTTPClientImpl) GetOverview(ctx context.Context, in *DashboardRequest, opts ...http.CallOption) (*OverviewPanel, error) {
	var out OverviewPanel
	pattern := "/api/v1/dashboards/overview"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDashboardsGetOverview))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DashboardsHTTPClientImpl) GetRules(ctx context.Context, in *DashboardRequest, opts ...http.CallOption) (*RulesPanel, error) {
	var out RulesPanel
	pattern := "/api/v1/dashboards/rules"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDashboardsGetRules))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DashboardsHTTPClientImpl) GetSources(ctx context.Context, in *DashboardRequest, opts ...http.CallOption) (*SourcesPanel, error) {
	var out SourcesPanel
	pattern := "/api/v1/dashboards/sources"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDashboardsGetSources))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}