# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/dashboards/disagreements:
        get:
            tags:
                - Dashboards
            operationId: Dashboards_GetDisagreements
            parameters:
                - name: timeRange.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: timeRange.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: interval
                  in: query
                  schema:
                    enum:
                        - BUCKET_INTERVAL_UNSPECIFIED
                        - BUCKET_INTERVAL_5M
                        - BUCKET_INTERVAL_1H
                        - BUCKET_INTERVAL_1D
                    type: string
                    format: enum
                - name: limit
                  in: query
                  description: Caps the rows a top-N panel returns.
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DisagreementsPanel'
...
    /api/v1/dashboards/overview:
        get:
            tags:
                - Dashboards
            operationId: Dashboards_GetOverview
            parameters:
                - name: timeRange.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: timeRange.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: interval
                  in: query
                  schema:
                    enum:
                        - BUCKET_INTERVAL_UNSPECIFIED
                        - BUCKET_INTERVAL_5M
                        - BUCKET_INTERVAL_1H
                        - BUCKET_INTERVAL_1D
                    type: string
                    format: enum
                - name: limit
                  in: query
                  description: Caps the rows a top-N panel returns.
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OverviewPanel'
    /api/v1/dashboards/rules:
        get:
            tags:
                - Dashboards
            operationId: Dashboards_GetRules
            parameters:
                - name: timeRange.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: timeRange.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: interval
                  in: query
                  schema:
                    enum:
                        - BUCKET_INTERVAL_UNSPECIFIED
                        - BUCKET_INTERVAL_5M
                        - BUCKET_INTERVAL_1H
                        - BUCKET_INTERVAL_1D
                    type: string
                    format: enum
                - name: limit
                  in: query
                  description: Caps the rows a top-N panel returns.
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RulesPanel'
    /api/v1/dashboards/sources:
        get:
            tags:
                - Dashboards
            operationId: Dashboards_GetSources
            parameters:
                - name: timeRange.from
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: timeRange.to
                  in: query
                  schema:
                    type: string
                    format: date-time
                - name: interval
                  in: query
                  schema:
                    enum:
                        - BUCKET_INTERVAL_UNSPECIFIED
                        - BUCKET_INTERVAL_5M
                        - BUCKET_INTERVAL_1H
                        - BUCKET_INTERVAL_1D
                    type: string
                    format: enum
                - name: limit
                  in: query
                  description: Caps the rows a top-N panel returns.
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SourcesPanel'
...
components:
    schemas:
...
        DisagreementPoint:
            type: object
            properties:
                bucket:
                    type: string
                    format: date-time
                kind:
                    enum:
                        - DISAGREEMENT_KIND_UNSPECIFIED
                        - DISAGREEMENT_KIND_NONE
                        - DISAGREEMENT_KIND_ALLOW_VS_BLOCK
                        - DISAGREEMENT_KIND_ALLOW_VS_CHALLENGE
                        - DISAGREEMENT_KIND_SCORE_CONFLICT
                    type: string
                    format: enum
                records:
                    type: string
                total:
                    type: string
                    description: |-
                        Every correlated request in the bucket. Carried alongside the numerator so a rate
                         can be computed from one row rather than by dividing figures fetched separately,
                         which is how two panels on one screen end up disagreeing.
        DisagreementsPanel:
            type: object
            properties:
                points:
                    type: array
                    items:
                        $ref: '#/components/schemas/DisagreementPoint'
                totalRecords:
                    type: string
                totalDisagreements:
                    type: string
...
        OverviewPanel:
            type: object
            properties:
                volume:
                    type: array
                    items:
                        $ref: '#/components/schemas/VolumePoint'
                verdicts:
                    type: array
                    items:
                        $ref: '#/components/schemas/VerdictPoint'
                totalEvents:
                    type: string
                    description: |-
                        Totals for the whole range, so a headline figure does not require the client to
                         re-sum the series and risk disagreeing with it.
                totalBlocked:
                    type: string
                totalChallenged:
                    type: string
...
        RuleCount:
            type: object
            properties:
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                ruleId:
                    type: string
                events:
                    type: string
        RulesPanel:
            type: object
            properties:
                rules:
                    type: array
                    items:
                        $ref: '#/components/schemas/RuleCount'
...
        SourceCount:
            type: object
            properties:
                clientIp:
                    type: string
                country:
                    type: string
                asn:
                    type: integer
                    format: uint32
                events:
                    type: string
                blocked:
                    type: string
                    description: |-
                        A subset of events, never a separate tally: a source showing more blocks than
                         requests is nonsense an analyst cannot act on.
        SourcesPanel:
            type: object
            properties:
                sources:
                    type: array
                    items:
                        $ref: '#/components/schemas/SourceCount'
...
        VerdictPoint:
            type: object
            properties:
                bucket:
                    type: string
                    format: date-time
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                verdict:
                    enum:
                        - VERDICT_UNSPECIFIED
                        - VERDICT_ALLOWED
                        - VERDICT_BLOCKED
                        - VERDICT_CHALLENGED
                        - VERDICT_RATE_LIMITED
                        - VERDICT_MONITORED
                        - VERDICT_UNKNOWN
                    type: string
                    format: enum
                events:
                    type: string
            description: VerdictPoint is one vendor's verdict count in one bucket.
...
        VolumePoint:
            type: object
            properties:
                bucket:
                    type: string
                    format: date-time
                vendor:
                    enum:
                        - VENDOR_UNSPECIFIED
                        - VENDOR_CLOUDFLARE
                        - VENDOR_F5
                        - VENDOR_DATADOME
                    type: string
                    format: enum
                events:
                    type: string
            description: VolumePoint is one vendor's traffic in one bucket.
...